Make Account#login nilable but assertive
There are a small handful of situations where the `login` property is
not provided, but it seems to be only in certain webhooks. Rather than
making this field generally nilable, making it assertive means you can
still count on it being there the vast majority of the time.

I really wish GitHub were more consistent in their serialization.