Fix Repos#list for the updated API
In 0c5d370d328a4865ae2916621889686c4043992e we removed the delegation
for HTTP methods to the client. It was attempting to be clever and it
turned out to be a bad idea. But it turns out we didn't update all of
the calls to explicitly call via the client. This commit fixes another
one.